+++ lib/crypt.dart
@@ -1,9 +1,7 @@
 // Copyright (c) 2015, Hoylen Sue. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-/// The crypt library.
-///
-/// One-way string hashing for salted passwords.
+/// One-way string hashing for salted passwords using the Unix crypt format.
 
 library crypt;