+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0
+
+- Match method and equality operator added.
+- Changed to represent the crypt as an object with a toString method.
+
 ## 0.0.1
 
 - Initial release.